coordinates, filters, and masks should not be in core module
