+++ src/cryojax/core/__init__.py
@@ -1,21 +1,6 @@
-from . import (
-    _coordinates,
-    _filter,
-    _mask,
-    _field,
-    _stacking,
-)
+from . import _field, _stacking
 
-from ._coordinates import *
-from ._filter import *
-from ._mask import *
 from ._field import *
 from ._stacking import *
 
-__all__ = (
-    _coordinates.__all__
-    + _filter.__all__
-    + _mask.__all__
-    + _field.__all__
-    + _stacking.__all__
-)
+__all__ = _field.__all__ + _stacking.__all__
