+++ src/cryojax/image/__init__.py
@@ -5,6 +5,8 @@ from . import (
     _spectrum,
     _normalize,
     _edges,
+    _filter,
+    _mask,
 )
 
 from ._fft import *
@@ -13,6 +15,8 @@ from ._average import *
 from ._spectrum import *
 from ._normalize import *
 from ._edges import *
+from ._filter import *
+from ._mask import *
 
 __all__ = (
     _fft.__all__
@@ -21,4 +25,6 @@ __all__ = (
     + _spectrum.__all__
     + _normalize.__all__
     + _edges.__all__
+    + _filter.__all__
+    + _mask.__all__
 )
